UserParamsとかをfloatからdoubleに変更

+++ Source/PluginProcessor.h
@@ -74,6 +74,6 @@ private:
     //==============================================================================
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(StereoPanAudioProcessor)
 
-    float UserParams[totalNumParam];
+    double UserParams[totalNumParam];
     juce::IIRFilter LPF[2];
 };
